refactor: separate error call into ButtonError component
