+++ src/components/Main/ButtonError/ButtonError.css
@@ -1,0 +1,3 @@
+.search-list__test-error {
+  margin-top: 10px;
+}
